% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toc_autoIndex.R
\name{toc_autoIndex}
\alias{toc_autoIndex}
\title{根据文件目录名称自动进行序号}
\usage{
toc_autoIndex(data = letters, prefix = "CH", sep = "-", digits = 2,
  symbol = "0")
}
\arguments{
\item{data}{原始数据}

\item{prefix}{前缀，如CH或第一章节}

\item{sep}{分隔符，一般使用-}

\item{digits}{序号位数，默认为2位}

\item{symbol}{补位符号，一般使用'0'}
}
\value{
返回字符串
}
\description{
根据文件目录名称自动进行序号
}
\examples{
toc_autoIndex(letters);
}
