% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toc_creator.R
\name{toc_creater}
\alias{toc_creater}
\title{在当前目录下创建目录}
\usage{
toc_creater(path_Names)
}
\arguments{
\item{path_Names}{字符串，向量}
}
\value{
返回true
}
\description{
在当前目录下创建目录
}
\examples{
toc_creator(letters)
}
